add: `colorscheme_exists` util func

+++ lua/huez-manager/api/colorscheme.lua
@@ -1,9 +1,9 @@
 local config = require("huez.config")
+local utils = require("huez-manager.utils")
 local M = {}
 
---- Reads and returns the persisted colorscheme as a `string` according to {path} or "huez-colorscheme" file.
----
---- If reading failed, returns `nil`
+--- Reads and returns the persisted colorscheme as a `string` according to the "huez-colorscheme" file.
+--- If reading failed, notify the user and return `nil`
 ---
 ---@param path? string
 ---@return string|nil
@@ -13,6 +13,7 @@ M.get = function(path)
   local file, _ = io.open(path, "r")
 
   if file == nil then
+    utils.notify("Huez: could not read the huez_colorscheme file.", "error")
     return nil
   end
 
@@ -22,9 +23,9 @@ M.get = function(path)
   return line
 end
 
---- Returns string[] where each item is the colorscheme name.
+--- Returns `string[]` where each item is the colorscheme name.
 ---
---- The array will contain all installed colorschemes if {exclude} is {}
+--- The array will contain all installed colorschemes if `Huez.Config.exclude` is an empty table
 ---
 ---@param exclude? string[]
 ---@return string[]
@@ -34,7 +35,6 @@ M.installed = function(exclude)
   local colorschemes = vim.fn.getcompletion("", "color", true)
   local manually_installed = {}
 
-  ---@type string
   for _, colorscheme in pairs(colorschemes) do
     if not vim.tbl_contains(exclude, colorscheme) then
       table.insert(manually_installed, colorscheme)
@@ -44,17 +44,25 @@ M.installed = function(exclude)
   return manually_installed
 end
 
---- Writes colorscheme to {path}, returns true if write was successful, false otherwise
+--- Writes a colorscheme to the huez colorscheme file if the colorscheme is valid
+--- and returns true if the file was written to successfully.
 ---
 ---@param colorscheme string
 ---@param path? string
 ---@return boolean
 M.save = function(colorscheme, path)
+  if not utils.colorscheme_exists(colorscheme) then
+    local msg = string.format("Huez: could not find colorscheme '%s'", colorscheme)
+    utils.notify(msg, "warn")
+    return false
+  end
+
   path = path or config.user.huez_colorscheme_file
 
   local file, _ = io.open(path, "w+")
 
   if file == nil then
+    utils.notify("Huez: could not write to the huez_colorscheme file.", "error")
     return false
   end
